requirements.txt
python-telegram-bot==21.5
openai==1.51.2
python-dotenv==1.0.1
aiohttp==3.10.5
